Fix the HyphenationCallback Typescript definition
According to the doc, the hyphenation callback is given the word as a
`string` and must return an `array` of `string` (`string[]` in TS).
That's how the features works according to my tests as well.

This commit fixes the type definition so as to accept the right type for
`word` (`string`).