Fix crash on `transform="translate(1)`
Providing one value to `translate()` produces valid SVG and is a common thing to see especially in minified SVGs.

When working with such SVG, I got a crash with a rather cryptic error message: "Error: unsupported number: undefined". Upon investigation, the lack of default `y` value turned out to be the cause.