feat!: bundle React to make it work with Next.js RSC environment
Fixes #2350
Fixes #2460

This proposal bundles React with @react-pdf/renderer.

The reason for that is Next.js replacing current React version with `18.3.0-canary-8c8ee9ee6-20231026`, which is not compatible with any stable version of react-reconciler available at the moment.

I realize that this is a huge change and I'm actually not convinced it's the right way to go, but if anyone would e.g. like to make a custom build on their own, this PR makes it possible to use @react-pdf/renderer in Next.js route handlers seamlessly.