Ensure zero advanceWidth for soft hyphens in both font and pdfkit
For the line breaking algorithm, soft hyphens should be considered to
have a width of zero, since they are never printed directly (they can
only lead to an inserted hyphen if at the end of a line).

The font package was already doing this correctly, but the pdfkit
package considered the soft hyphen to be the same as a normal hyphen
with an advanceWidth of 333 in Helvetica. Without this change, in some
edge cases the pdfkit would break apart lines already broken apart by
the line breaking algorithm in textkit.

Added tests for both packages to make sure they remain compatible in the
future.