Always return both height and width when measuring text
Fixes #2974
Returning an undefined width translated into a NaN width when passing
the measurement result back to yoga. This NaN width would then be
converted to 0 and a warning as follows would be emitted:
Measure function returned an invalid dimension to Yoga: [width=nan,
height=21.779297]
The warnings in yoga's javascript bindings are stored into a long
array buffer. If a document is very large (> 100 pages) and contains
a lot of text nodes, this can produce millions of yoga warnings,
inflating this array buffer, until at some point the browser refuses to
push another character onto the end of the buffer. This results in the
error `RangeError: Invalid array length at Array.push ()`