Try harder to avoid infinite loops while page splitting
When a node is already at the top of the page, it does not help to break
it to the next page. We knew that already, but the implementation of
this knowledge has been flawed. Apparently, page paddings, page borders,
fixed elements and possibly other things can cause the previous
position-based detection method to not work correctly.

This commit replaces the flawed implementation with a hopefully more
solid one: When there are no non-fixed nodes before the current one on
the page, we can assume that the current node will not be placed higher
up on the page even if we were to break it to the next page.

Fixes #2884
Fixes #2895
Fixes further parts of #2974